GRAL-1759 ItemController
