+++ lib/index.js
@@ -15,6 +15,7 @@ const ActivityTypesController = require('./Controllers/ActivityTypesController')
 const CurrenciesController = require('./Controllers/CurrenciesController');
 const DealFieldsController = require('./Controllers/DealFieldsController');
 const DealsController = require('./Controllers/DealsController');
+const ItemController = require('./Controllers/ItemController');
 const FilesController = require('./Controllers/FilesController');
 const FiltersController = require('./Controllers/FiltersController');
 const GlobalMessagesController = require('./Controllers/GlobalMessagesController');
@@ -269,6 +270,7 @@ const initializer = {
     CurrenciesController,
     DealFieldsController,
     DealsController,
+    ItemController,
     FilesController,
     FiltersController,
     GlobalMessagesController,
